Migrate to Paket
Migrated to using Paket for everything. This requires keeping paket.exe. I was also having trouble building the tests at net461. I bumped it to netcoreapp2.0. Most of the dependencies have been migrated to netstandard2.0. There are a few that are WIP, so until they are migrated this is a WIP.